backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic[email]==2.8.0
asyncpg==0.29.0
python-dotenv==1.0.1
openai==1.51.0
httpx==0.27.0
twilio==9.3.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
aiokafka==0.11.0
watchdog==5.0.0
schedule==1.2.2
playwright==1.47.0
pytest==8.3.0
pytest-asyncio==0.24.0
